feat: compute Event.subscriberCount from Subscription data
Subscriber count was previously hardcoded to 0 since Subscription
didn't exist yet. Now derived at read time via SubscriptionRepository
(single-event count for getById, grouped batch count for getAll to
avoid N+1), never stored/incremented, so it can't drift from the
truth the way a denormalized counter could.

EventService.getById is restored to return the Event entity (needed
by SubscriptionService, which composes it) with a new getResponseById
for the count-enriched DTO path.

+++ src/main/java/com/example/relay/event/api/dto/EventResponseDto.java
@@ -8,5 +8,5 @@ public record EventResponseDto(
     String name,
     UUID appId,
     Instant createdAt,
-    Integer subscriberCount
+    long subscriberCount
 ) {}
